+++ doc/Changelog.md
@@ -1,4 +1,6 @@
+* give the user the chance to download the album, either with or without its subalbums (shortcut _D_ added)
 * added 3-9 enlargments by shortcut
+
 
 ### version v4.0.4 (Nov 11, 2019)